19:47 - 15 Oct 2016

+++ app/src/main/java/com/minhnpa/coderschool/flicks/model/Movie.java
@@ -40,6 +40,6 @@ public class Movie {
     }
 
     public int getVoteAverage() {
-        return Integer.parseInt(voteAverage);
+        return (int) Double.parseDouble(voteAverage);
     }
 }
